requirements.txt
# Core dependencies for Nougat OCR worker
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
supabase>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
